fix(buck2_common): be less aggressive about HTTP retry warnings
Summary: When downloading from sites like crates.io, we can pretty easily get
rate-limited when a lot of our crates are uncached. Don't be so aggressive about
warning; most small backoffs won't be too noticeable.

As a note, we _don't_ emit the error message we get from the callstack here in
the message; the logic is that it's already a retriable error, so it doesn't
make much sense to spam an error that we'd probably get again anyway, and
that we ultimately are going to re-attempt. This should just make the UX a
little more polished.

Also emit the domain name we're hitting problems with.

Test Plan: Change the new warning conditional from 30 seconds to zero seconds.
Build buck2 with buck2. Observe the new error message, which helpfully points
out we're connecting to crates.io. Change the conditional back to 30 seconds.
Rebuild buck2 with buck2. Note that no warnings appear, because the exponential
backoff only results in a few small 2-second retries.

GitHub Issue: Fixes #316

Signed-off-by: Austin Seipp <aseipp@pobox.com>
Change-Id: Ixoxuusxlkkrspllqqvukzrumnrytmpko