Guarantee exec deps are for the right execution platform
Summary:
In the workaround from D46750398 (before `exec_dep` was usable in anon targets), I made `http_archive` have an ordinary `dep` on "prelude//http_archive/tools:exec_deps", which had an `exec_dep` on the necessary tools which http_archive needs to execute.

This worked in practice most of the time but was not guaranteed to do so, [as flagged by cjhopman](https://fb.workplace.com/groups/buck2eng/posts/3201623793468432/?comment_id=3202019773428834&reply_comment_id=3202433056720839).

> No, you can't propagate exec_deps in that way.

There was no guarantee that the execution platform of "prelude//http_archive/tools:exec_deps" would be the same as the execution platform of the http_archive downstream of it. If they were different, the wrong tool would be invoked during the execution of http_archive.

This diff fixes that, based on wendy728's implementation of `exec_dep` support for anon target ({D46973087}).

Now http_archive has an `exec_dep` on "prelude//http_archive/tools:exec_deps", which has an ordinary `dep` on all the tools. :exec_deps is configured for the execution platform of the downstream http_archive, and therefore so are its `deps`.

Reviewed By: zertosh

Differential Revision: D46978840

fbshipit-source-id: 00d64c29bdd76668ce3a3f9297d1576fd99a04fe