buck2 OCaml prelude: ignore mlis not listed in srcs
Summary:
## What

ignore mlis not listed in `srcs`.

## How

1-line change:
Give the OCaml compiler a nonsense extension instead of `.mli` to use for looking for an mli when we know there is no corresponding .ml for a .mli.

## Why

Ignoring mlis that are not in `deps` enables us to avoid the following scenario:

`build -> add .mli for a .ml -> build (OK) -> clean -> build (fail: "Could not find the cmi file ...")`

That scenario could previously happen (D47058043, see comments) because the incremental build would pick up an old interface generated from a .ml but a clean build would find no interface file.

Reviewed By: vsiles, shayne-fletcher

Differential Revision: D47091314

fbshipit-source-id: 1cdb25fa83ef8a36624ef7707715937db64b0654