Defer notifying completion until TaskHandle is dropped
Summary:
Change DiceTasks to only notify completion or cancellation when the entire task is dropped.
Previously, finished tasks were allowed to notify completion earlier than when a task is cancelled. This simplifies code so that no matter completed or cancelled, we always only notify on completion.

This has a nice properly that after D46744411, the drop is also tied to the completion of the explicitly cancellable future, which drops on completion, so now all the task completion and cancellations and drops all happen together.

Reviewed By: krallin

Differential Revision: D46744409

fbshipit-source-id: f61e064e74a2bd9212737b86f0866b32a1562318