Acquire prevent cancellation before the future starts
Summary:
There's a rare race where right after we spawn the future, before we manage to execute the actual critical section in the future, we cancel the task already. Then we end up never waiting for the cancelled tasks.

This prevents cancellation immediately in a sync block to ensure that the future is not cancelled in the beginning.

also moves more disabling cancellation logic into the states since we use the new cancellation api, which is easy

Reviewed By: krallin

Differential Revision: D46745349

fbshipit-source-id: be7ddd02b3d3b0a45f0fc1549eca499e5e85f9db