buck2: emit a full snapshot in pre-initialization
Summary:
Because we can, but also because that's the right thing to do. We have a bunch
of things that look at a delta of the first snapshot vs. the last one, but
right now we emit a snapshot that's mostly empty as the first snapshot.

We could fix this by having some fields properly expressed as optional, but in
reality we don't need any of this, because everything we want to collect is in
fact already available at the time we capture our first snapshot, so we might
as well emit the full thing.

This is what this does. If we want to add things available later, we can do
that by simply mutating the collector later to add those fields.

This actually puts the rest of this stack together and ensures that our HTTP
and RE downloads are now properly using a non-zero data point as their
reference (before, they'd use a datapoint that was zero so they would
always show the daemon lifetime download amount, which is bad).

Reviewed By: get9

Differential Revision: D47263599

fbshipit-source-id: 12744f19a2a2c6f2ea8ca0e125c8dcdc310eaff7