buck2: remove stats from the `EventSink` trait
Summary:
We try to pretend that this works for all `EventSink` but it only works for
Scribe, so let's stop lying to ourselves and express is it as such.

This also shows why using the daemon's event sink for stats in the
SnapshotCollector is the way to go: there is no other sink than this one that
provides stats.

Reviewed By: get9

Differential Revision: D47263600

fbshipit-source-id: 0cbca62a5acc6cde890475eabafbf312671d88cd