remote_execution: expose network stats without a client
Summary:
This actually just queries a singleton, so we might as well let the caller
access it without a client.

If the stats were per-client, the current implementation would make sense, but
there seems to be little appetite to do that (it's been like this for a while,
and the only consumer of this — Buck — does not care), but they're not, so
considering instantiating a RE client is expensive but accessing a singleton
is not, let's just remove the indirection through a RE client.

Reviewed By: get9

Differential Revision: D47263605

fbshipit-source-id: 7ae835bae967662670dc1c29a77772b96b6b40c0