Use full .jar files when generating class_to_source_map
Summary: In general, we use ABIs in our rules because they are cheaper. However, in the case of tests, we need all the full jars anyway as that's what goes on the classpath, and relying on ABIs means creating some ABIs that we wouldn't create otherwise (and don't create in buck1). So, let's just use the full jars.

Reviewed By: krallin

Differential Revision: D47213840

fbshipit-source-id: f5add2665788975615e52f027f79185b81ffc5ce