Fallback to regular url if no vpnless_url attribute
Summary:
When running in vpnless mode, let's prefer to use the `vpnless_url` attribute but fallback to the regular url if it's not specified.

This, plus some upcoming changes to x2pagent that allow routing all urls through the proxy will enable a better UX for people testing out vpnless. Instead of having buck bail because it had no vpnless_url attribute, we'll fall back to the regular URL which should work if you reconnect to the vpn.

Reviewed By: krallin

Differential Revision: D47266212

fbshipit-source-id: 484061e244e6c1674214e88e42f843c038e58a11