Where possible, prefer unix domain socket over http proxy port
Summary:
According to johansglock and ngoyal, it's better to use the UDS proxy exposed by x2pagent rather than the http proxy. Let's do that.

This only works currently for unix-likes because the [hyper-unix-connector](https://docs.rs/hyper-unix-connector/latest/hyper_unix_connector/) (which is what we use to connect to the x2pagent UDS) relies on tokio's unix connector client which doesn't support windows.

For windows, we'll still prefer to use the http proxy server.

Reviewed By: krallin

Differential Revision: D47266211

fbshipit-source-id: 427829f3e360af7a2fe4e4907d7f87c62dc7a4ae