Restart daemon when vpnless is toggled
Summary:
It's desirable to restart the daemon (so we switch to the appropriate client) when `buck2.allow_vpnless` changes. This config controls what kind of client we use - e.g. if we use a client that talks to x2pagent vs one that does not.

This causes buck to restart if the value of this config changes between invocations.

Reviewed By: krallin

Differential Revision: D47266210

fbshipit-source-id: b3503aea9e0e07231e94eba33e02429b5992ab5a