Always use arm64 as the primary platform
Summary:
When we build Android binaries, we do a "split" transition at the binary level, such that all the deps are given a new configuration that includes the cpu that we want to build native libraries for (and if it's a multi-cpu build, they are given multiple such configurations).

We want to build the non-native libraries with the same configuration regardless of which cpus are specified, so that we can re-use the same actions for build that are targeted at different CPUs.

Therefore, always build the non-native libraries using the arm64 platform (which is the most common platform that we will build for, by far).

We don't build them with no cpu at all because in the most common case (building for the arm64 platform) we want to optimize for the best possible memory usage.

Reviewed By: krallin

Differential Revision: D44956000

fbshipit-source-id: 53662259b57f2de1e8e145391fc405b05a3196b5