rust-project should use crate_root exactly
Summary:
If a crate has an explicit `crate_root` specified, use it directly.

The fallback logic looks for the first file in that ends with `lib.rs`, `mod.rs` or `crate_name.rs`. This was inappropriate for an explicit `crate_root`.

This previous behaviour would break if the user specified their root was `foo.rs` and `abc/foo.rs` also existed. We would treat `abc/foo.rs` as the crate root and IDE services break.

Reviewed By: zertosh, shayne-fletcher

Differential Revision: D47306751

fbshipit-source-id: 19a2c6444b20c1d216718feb7a13ff117644e8a2