workers; block worker execution if over worker configured thread limit
Summary:
JVM based workers can face a number of issues when a single worker runs too many threads in parallel on machines with large numbers of cores:
- Non-deterministic outputs (concurrency bugs)
- Annotation processor crashes (concurrency bugs)
- Excessive garbage collection
- File descriptor limits ("Too many files open")

These issues are independent of machine capacity and other concurrent build actions so we should limit this per worker.

We should try to address the concurrency bugs, if possible, but this is still useful for the sake of more predictable memory usage and performance.

This could be implemented in the workers themselves, but supporting it in buck is better so that queuing isn't counted as `worker_execution` and to avoid a possible performance penalty if a worker is cancelled while queued.

Reviewed By: krallin

Differential Revision: D47281221

fbshipit-source-id: bc8a3b5b69e43f3f113b9ec69083f558f86e2374