Implement within_view (but not enable yet)
Summary:
`within_view` kind of works, but:
* buck2 `within_view` works differently (before select expansion, while buck1 seems to do it after)
* probably different types of dependencies are checked/not checked differently
* macros work different with buck1/buck2

So when trying this diff unconditionally I encountered problems like this:

P777918216

(the target mentioned in error message does not even exists in buck1)

So now there's an option (on by default, turned off in the previous diff D47012711) to enable `within_view` checks.

Let's land this diff and then fix the repo, and then enable the option.

Reviewed By: krallin

Differential Revision: D47011628

fbshipit-source-id: 32cdcd8e5e08e003872bfb292628fd8f0ff020a9