Log a single worker init span per worker
Summary:
Right now an `InitializeWorker` span is logged for every action that uses a worker, even if the worker has already been initialized.

If the worker has already been initialized, there's no need for a span, and if another action has created the initialization future, it's not necessary to log the worker initialization args again, so this shows a `WorkerWait` span instead.

This removes a lot of noise from logs, but also allows using `what-ran` to show the number of workers initialized during a build (next diff).

Renamed `InitializeWorker` to `WorkerInit` for consistency with `WorkerExecute` and `WorkerInitCommand`.

Reviewed By: krallin

Differential Revision: D47202169

fbshipit-source-id: f8425090a57019d26482ce5dc317949df3961bb9