fix(cargo): update bumpalo 3.8.0 -> 3.11.1 (#333)
Summary:
Fixes an inconsistency between the version used for allocative, Cargo.toml, and the shim Cargo.toml.

I found this while doing some experiments with https://github.com/facebook/buck2/issues/308, and figured it might be nice to have.

There is also an inconsistency in the `starlark-rust` crate (which uses 3.8), but I'm not fixing that here unless asked because I believe those fixes need to go to the starlark repository.

Pull Request resolved: https://github.com/facebook/buck2/pull/333

Reviewed By: scottcao

Differential Revision: D47337433

Pulled By: themarwhal

fbshipit-source-id: 61d9d3f891d959bcd4aa44b67435ac71a528ad17