starlark: add support for lexing f-strings
Summary:
Like it says in the title. This doesn't do much besides just extracting the
string this contains, and right now it does nothing because the parser does not
handle this at all.

I'm not checking the dialect allows those yet since I'm not sure if it's best
to do it in the lexer or the parser yet.

Reviewed By: stepancheg

Differential Revision: D47291008

fbshipit-source-id: 491df40daa6954fb9ebb7c0ec19e3166ec0dff5b