fix grep not found issue for tag check logic in GHA
Summary:
I thought I fixed this, but I hadn't. grep returns 1 if the string is not found (https://github.com/facebook/buck2/actions/runs/5521152524/jobs/10069658332)

Break apart the curl bit (shouldn't error) from the actual grep bit (can return 1 if it's not found) so that I can tack on a `|| true`. As far as I know, there isn't a grep flag that makes it so that it doesn't error on the not found case.

Reviewed By: wendy728

Differential Revision: D47368275

fbshipit-source-id: 28458370a1edc3cf4b5e8b1a28edae3c5ce62819