Update 'get_actual_link_style' to incorporate PicBehavior
Summary:
Now that we have a `PicBehavior` enum, lets update the function we use within `prelude` to decide the `link_style` for most targets.

To make this easier to review, I'm made a default argument so that you can simply review the logic. In the next diff I'll make it required.

Reviewed By: chatura-atapattu

Differential Revision: D47198672

fbshipit-source-id: 3a398443bc63857f2b47e4451a1fd593fd10a39e