Add a new type of update request for matching deps
Summary:
There's two types of updates to state

1. when we recomputed the value,
2. when we determined that the deps are unchanged and hence can reuse the value.

In the former case, we should check for equality to determine if the node can be reused.
In the second, we need to verify that the version for which we checked deps are consistent with what we currently store in the state, if so, reuse it. We shouldn't use equality because some keys' values don't support equality (i.e they are always false), but they still want to be resurrected if their deps are unchanged.
Prior to this diff, we used equality incorrectly. Now with this diff, we stop using equality and resurrect the node.

Reviewed By: stepancheg

Differential Revision: D46955223

fbshipit-source-id: c6d93b8ad062c9fa647ef39713a4dd14b420cc5b