Keep separate tset for external debug info
Summary:
Currently, each `LinkInfo` holds an `ExternalDebugInfoTSet`, to track all
external debug info associated with the items in link info.  When we perform
a link, we need to expand/traverse the `LinkInfo` tset that `LinkArg` wraps
to collect all nested `ExternalDebugInfoTSet`s and then wrap them in another
tset.

For builds w/ lots of large links, this is becoming noticeably expensive (e.g.
20% of some builds).  This diffs switches to building up this tset as we go,
so that it's essentially available for free for all dependent links.

Reviewed By: chatura-atapattu

Differential Revision: D47269809

fbshipit-source-id: 39506e51e10dd6e10b278705fad7fc62cee2ba30