Avoid locks in dicepromise and cancellation observer poll ready
Summary:
We conditionally check the wakers and remove the waker of the same future when the future is dropped.
However, when the future is completed via a poll, it will still be dropped and trigger the waker removal, increasing contention on locks. This reworks the code so that we don't lock when poll is ready.
The wakers list will have been cleared anyways by the original future that completed the task.

Reviewed By: krallin

Differential Revision: D47305047

fbshipit-source-id: d9c23b1ce44e3d0c774bb4ab3f6e13d59eb60219