Allow specifying used-classes jar via buckconfig. Take 2
Summary:
This diff is a copy of a Take 1, but now the feature is turned off by default with a fallback to a previous implementation

Buck 1 & 2 will consume `track_class_usage_plugin`.
`track_class_usage_plugin` is a dep-tracker plugin that was previously a part of the buck codebase.

Introducing a `DepTrackerConfig`, a dedicated config for all things related to tracking class dependencies

Reviewed By: natestedman

Differential Revision: D47417998

fbshipit-source-id: 8cf2d40ea4beb5a08c2956c2eaa077d2311d3560