Move BxlSafeDiceComputations to buck2_build_api
Summary:
This stack will make the `bxl_ctx` available within the dynamic lambda. The end state is that if we detect that the dynamic lambda was created by BXL, we will create a `bxl_ctx` instead of an `actions` (`AnalysisActions`) context.

We will create an object that has some, but not all of the `BxlContext`'s functionalities. For example, I think we should be able to run queries, builds, filesystem checks, but `ctx.output.print` and `ctx.output.ensure` probably shouldn't be supported for simplicity. We are able to run multiple dynamic outputs concurrently, and also in a nested fashion, so we run into some determinism issues if we allow for `ctx.output.print` and `ctx.output.ensure`. In addition, if a `bxl_actions()` object is created inside the dynamic lambda, it will always inherit the execution platform resolution from the parent.

I thought about somehow using late binding to be able to create a `BxlContext` (which lives in `buck2_bxl`) as a `Value<'v>` within the `buck2_build_api` crate, but since `Value<'v>` is not send or sync, and `BxlContext` has various dependencies that aren't available within `buck2_build_api`, it would require a lot of refactoring, and most likely involve moving almost all of `buck2_bxl` into `buck2_build_api`. The plan highlighted below seems cleaner to me, but open to suggestions.

Plan:
* Implement a `BxlDynamicContext`, where the starlark type is still `bxl_ctx` (I think this works?). This is only accessible from within the dynamic lambda. This struct will live inside `buck2_build_api` so that we can access it when we actually evaluate the dynamic lambda. [Code pointer](https://www.internalfb.com/code/fbsource/[0e3986938f9cdb5ffee57b1b4c5271d8e6c76bd8]/fbcode/buck2/app/buck2_build_api/src/dynamic/deferred.rs?lines=263-289) for where it will be instantiated
* Move `BxlSafeDiceComputations` from `buck2_bxl` to `buck2_build_api`. This is so that we can access this object within the `BxlDynamicContext` object
* Add a `BxlContextData`, which includes common/relevant fields to be shared between `BxlDynamicContext` and `BxlContext`, including things like cell resolver, cell name, `BxlSafeDiceComputations`, target alias resolver, and global target platform (maybe - see open questions). Then, replace the fields in `BxlContext` with this `BxlContextData` object. This will involve a lot of refactoring because several starlark objects (ex: `StarlarkCqueryCtx`, `StarlarkUqueryCtx`) have a reference to `BxlContext`.
* The user who wants this specified that `configured_targets()`, `analysis()`, and `resolved_attrs_lazy()` are the most important functions, so I will prioritize these as I slowly add functionalities to `BxlDynamicContext`.

Open questions:
* Not sure if we want the `bxl_ctx` within the dynamic lambda to inherit the global target platform from the parent BXL, or it should be None?

Anyhow, I'm putting up the first diff with a detailed summary to see how people feel about this approach.

Reviewed By: bobyangyf

Differential Revision: D47409520

fbshipit-source-id: d240769a2e17445c356bf5d020ca6cc5f10a57bb