Export split debug handling (take 3)
Summary:
When using split-debug modes and LTO on macos or linux, the link step will emit
debug info objects that we need to track via external build info.

Export this directory in the `LinkedObject` struct, so that we can access it from
the caller.

This attempt fixes an issue in fbcode split-dwarf thinLTO builds where
we didn't properly bind the DWO output dir.

Reviewed By: christycylee

Differential Revision: D47413573

fbshipit-source-id: e5b1089ca4a0839e309767472a774683cbec1978