buck2: paranoid: ignore prefer_local & prefer_remote
Summary:
When those flags are set, we disable hybrid execution. We don't want that in
paranoid mode (see the bottom diff), where we'd rather just let everything race.

To do that, we introduce a new executor preference that amounts to deleting
preferences, and use it when configuring the hybrid executor for paranoid mode.

Reviewed By: get9

Differential Revision: D47406607

fbshipit-source-id: f32a94b181745d7c96f7815467b9b441870f7f4f