buck2: paranoid: add a StackedExecutor
Summary:
I'm going to use this to support racing action cache queries and local
execution. To do this, we'll move action cache queries into an executor that
queries the cache then executes locally. That whole process will run in parallel in local execution.

Note that we won't do this by default. The goal is to use this for the
paranoid mode only (see the bottom diff for context).

Reviewed By: get9

Differential Revision: D47406606

fbshipit-source-id: dd75c34df3e08db964252e007b22635b8f629af9