buck2: improve an error message
Summary:
If we cancel local then try to download from RE and fail, we'll favor showing
the local executor's error, but that will not be very helpful because it'll
be reported as "Cancelled".

The more useful error to show is the RE executor's, which would have taken
the claim (causing the cancellation) then failed. So, let's just do that.

Now, obviously, the error output will leave open the question of "but *why did
this not retry locally*?", but really this behavior is a) subtle and b) rare,
there's just not going to be a good way to succinctly explain what went
wrong.

Reviewed By: themarwhal

Differential Revision: D47429620

fbshipit-source-id: 0f325e90a9b4d606702da897bcd294f60af5d254