buck2: paranoid: move the action cache queries into the RE client
Summary:
In paranoid mode (see the bottom diff), we want to do do cache queries as part
of the RE client. The normal cache queries we do before hitting the executor
become a noop (this will also eventually mean we don't get e.g. "dep files from
stable", but that's also a feature we want to turn off).

Reviewed By: get9

Differential Revision: D47406605

fbshipit-source-id: 618b8ca8af07e27ad6931c67868a642fc6de419d