chore(nix): use mold for linking inside 'nix develop' shell
Summary: when using Linux, and developing with `nix develop` or `direnv`, go
ahead and enable mold as the default linker.

We modify `RUSTFLAGS` in the shell environment to do this so that we don't have
to fiddle with the internal `.cargo/config.toml` file; and because it's only
for development anyway, this is strictly used only while developing, not while
running `nix build`.

On my Ryzen 5600X, versus the default linker settings, this takes the link time
for the buck2 executable from 32s to 3s.

Signed-off-by: Austin Seipp <aseipp@pobox.com>
Change-Id: I66818b3f45ce07119a42afac7320df78