Don't allow trailing path separator in $OUT
Summary:
Initial report here:
https://fb.workplace.com/groups/930797200910874/permalink/1250269588963632/

If `out_env` is the empty string, an extra path separator was getting appended to the `OUT` genrule environment variable. All the genrules are inconsistent with how they handle quoting, but some use `"%OUT%"`. That results in a `\"` at the end of what was intended to be a quoting argument, which escapes the double quote mark, and appends other arguments to the `OUT` argument. Bad news.

Solution: if `out_env` is empty, don't append it.

Reviewed By: akrieger

Differential Revision: D46960280

fbshipit-source-id: e6c01b7cace8f64d07eb410ad01379a0e8e08351