buck2: stdin: retry EINTR
Summary:
On Mac you can get this if you CTRL+Z Buck2 while it's running then resume it.
I'm not sure why the signal gets delivered to this thread there and not on
Linux, but here it's clearly not wrong to retry EINTR so we might as well just
do so.

Reviewed By: lmvasquezg

Differential Revision: D47511075

fbshipit-source-id: 6b9dba6f1ad75a40342531a9e2b22ca26e07efb7