Override execution strategy and caching when setting BUCK_OFFLINE_BUILD
Summary:
For offline builds, we need to always override certain flags (specifically `--local-only` and `--no-remote-cache`) because this build is done in a fully isolated, no-network container.

We need to override this at the client level - as opposed to configuration level - for a few reasons:
* Some tooling in the container calls buck (notably, fbpkg-build).
* Updating the configuration via e.g. `execution_platform` changes configuration hashes which invalidates pre-cached network action outputs inside the container.

As a result - let's resort to an environment variable. We already have shell script wrappers laid down in the container (prior diffs) that have to define additional environment configuration for e.g. dotslash, so we can define this variable there as well.

Notably, we also want to make `--local-only` and `--remote-only` conflict - you obviously can't pick a local-only _and_ remote-only executor.

Reviewed By: krallin

Differential Revision: D47479609

fbshipit-source-id: be4fcda8265939da9925845c314da286b3c31399