Prefer reporting incompatibilities over errors
Summary:
The amended test from this diff looks like this:

```
prelude.export_file(
    name = "windows-only",
    src = select({
        # In test we run the query with target platform linux,
        # this select does not have DEFAULT,
        # but it should not fail because it should not be evaluated at all,
        # because of `compatible_with` check.
        ":windows": "TARGETS.v2",
    }),
    compatible_with = [":windows"],
)

prelude.export_file(
    name = "windows-only-undeclared",
    src = select({
        # This rule is only compatible with windows, however we don't declare that
        ":windows": "TARGETS.v2",
    }),
)

two_dep_rule(
    name = "windows-only-deps",
    a = ":windows-only-undeclared",
    b = ":windows-only",
)
```

Before this diff, cquery with linux as the platform fails on `:windows-only-deps`, complaining that `src` can't be configured on `:windows-only-undeclared`. However, this only occurs because we happen to attempt to configure `:windows-only-undeclared` before reporting the incompatibility from `:windows-only`. Indeed, if the attrs are declared in the opposite order then this test passes.

Having this depend on the attr declaration order is weird. We should either always prefer to report the incompatibility or always prefer to report the error. This diff changes the behavior in favor of always reporting the incompatibility, and only reporting an error if there was no incompatibility.

I can see good arguments that this is actually the wrong behavior to choose. The `:windows-only-undeclared` target in the test really should declare its compatibility. Unfortunately, opting for the opposite behavior breaks the repo (buck2 doesn't even bootstrap). That would have to be fixed first.

Reviewed By: krallin

Differential Revision: D46910954

fbshipit-source-id: 9dbd4c2b73c55e365ce41fe9d9709a4a8aebf4a4