Fail on non 200 curl exit code and retry curl a few times
Summary:
Noticed today that there was a third release tag that went out yesterday. (https://github.com/facebook/buck2/actions/runs/5590979823/jobs/10222098201)
This was because the curl command failed due to a transient failure, but we proceeded to treat the result as an empty string, concluding there were no tags yet.

So properly fail on curl failures, but also add some retries so that we don't fail every time we experience a transient curl issue.

Reviewed By: stepancheg

Differential Revision: D47594344

fbshipit-source-id: da20087f1557ea2230ed2ca7005163f108fac90f