buck2: omit kq_dispatch logspew in console
Summary:
See T155741660 for some context. Those messages seem to be triggered by a bug
in Thrift / Folly where it closes a FD but then does not immediately removes it from
its list of FDs to poll, which results in this error.

After finding that root-cause, I went looking for help in debugging this but so far
that's not been too successful:

https://fb.prod.workplace.com/groups/560979627394613/permalink/2719867511505803/

In time, I might revisit this, but for now I'm confident enough that this
warning is innocuous, so I'm just updating our console to stop printing it.
We'll still log it to the daemon's stderr.

Reviewed By: christolliday

Differential Revision: D47590586

fbshipit-source-id: c0b1771acfa6b6a3acda585b8de2bf9a3911f54c