buck2: fill in the `metadata` attribute
Summary:
Like it says in the title, this makes the attribute actually hold metadata
(but it cannot be constructed yet).

I don't really love the current design of using `serde_json::Value`. I'm doing
this to be consistent with package values, but I think we should change that.

There are a few reasons:

- Hashing `serde_json::Value` is annoying.
- We can't intern any nested keys.
- We don't really control the representation (e.g. `serde_json::Value` uses
  either BTreeMap or IndexMap for its maps, depending on features, and we
  probably want to make this a more explicit decision, especially within Meta
  where features on third-party crates are a bit of a pain).

That said, for now this provides a reasonable stand in before I go and refactor
both this and package values, which can be done without changing the API.

Reviewed By: stepancheg

Differential Revision: D47586464

fbshipit-source-id: 807d811426017c0e5f9f4a43ad2c771b913ec413