buck2: BuckSpawner: explicitly capture a runtime
Summary:
Like it says in the title. There is a big of a change here in that this diff
makes us capture the runtime when instantiating the BuckSpawner, whereas before
we did so when spawning but we only have one, so that doesn't matter (but that
changes later in this stack).

Reviewed By: bobyangyf

Differential Revision: D42846351

fbshipit-source-id: 7da80c6b30376d81efeeae6abd7ec6f10fde08d9