buck2: expose the daemon's digest config to rules
Summary:
It's already basically there. I need to expose this because there is one place
in http_archive where we try to guess whether the daemon will be able to defer
a given digest.

I put this on `ctx.actions` since we a) already have the digest config there and
b) the places where you might care about this are places where you should
have `ctx.actions`.

Reviewed By: themarwhal

Differential Revision: D47665045

fbshipit-source-id: 92f52a5ebb2e4385bfb56f9cc201b6c70343b331