Inline several unpack calls
Summary:
Calls like `AValueDyn::is_str` do not require instance. Use vtable directly.

**Context**: sometimes we need `me` `Value` in functions. Thinking how to do this more efficiently. Currently when it is used, like in `invoke`, we essentially pass the pointer twice:

https://www.internalfb.com/code/fbsource/[20324261d7035c009fef83ce765abcb1c6e01d00]/fbcode/buck2/starlark-rust/starlark/src/values/traits.rs?lines=391-393

This is overhead already, and if we do it for all operations, it may be larger perf overhead. Ideally we should only pass `Value` to virtual calls and unpack it inside the function. And to do that, we need to change `AValueDyn`, which currently contains raw data pointer, but may contain the `Value`:

https://www.internalfb.com/code/fbsource/[20324261d7035c009fef83ce765abcb1c6e01d00]/fbcode/buck2/starlark-rust/starlark/src/values/layout/vtable.rs?lines=241-245

**More context**: for operations like `str | int` we need to pass `me` to `StarlarkValue::bit_or`.

Reviewed By: JakobDegen

Differential Revision: D47685891

fbshipit-source-id: d4d6241ed81fec710fb0187753f2c81328da52cb