buck2/toolchains: run link.exe through msvc_tools
Summary:
Use the same approach as in D47702231 to find `link.exe`. It already finds correct `LIBPATH`.

With this, none of MSVC tools need to be added to `PATH`.

This should fix out Github Actions job which is currently failing because it can't compute `LIBPATH`: https://github.com/facebook/buck2/actions/runs/5643015426/job/15284016654

Reviewed By: zertosh

Differential Revision: D47722172

fbshipit-source-id: a47b87ea54ebd9a902e4687a4367fe510bb76841