buck2: upload logs: soft_error! on error
Summary:
**why this**
We currently have no way of knowing what happened when upload success rate goes low

**this diff**

use soft_error! mechanism to extract that information

> // Throw a "soft_error" i.e. one that is destined to become a hard error in the near future.

This diff is making that comment be inaccurate. If we truly want to keep it that way, then I offer myself to create an equivalent macro that will do the same, just with a different overarching goal

**other alternatives**

We came up with T158169647 to solve this same issue, yet this diff approach is cheaper to implement and more effective. (Thanks to Sean for the idea!)

Reviewed By: krallin

Differential Revision: D47527440

fbshipit-source-id: e91b11163f0aa995ccedf18d65feff7c93b11f1d