Runtime version of int | str (v2)
Summary:
Alternative to D47694041: `bit_or` for types is generated for proc macro instead of being handled at `Value::bit_or`.

`StarlarkValue` implementation is considered "type" (i.e. `bit_or` is generated) if the implementation implements `eval_type` function.

`|` is not generated for types like `[str]` which are deprecated (proper expression for this type is `list[str]`). So

```
[str] | [int]
```

works (for now) in type expression, but does not work as regular expression. And

```
list[str] | list[int]
```

works in any context.

Reviewed By: krallin

Differential Revision: D47725494

fbshipit-source-id: 86409e0f053f2917d8ae03fd9f10d87e531a0174