buck2: move re_queue_time to `CommandExecutionKind`
Summary:
Two reasons for this:

- It is better if our internal types and data types have the same shape, and
  right now internally this comes from `timing` but goes into the thing that
  notionally represents `CommandExecutionKind`
- Only one case has this. The value we use for this in action cache hits is
  completely meaningless right now (and there is a comment claiming it is
  something it absolutely is not)

Note that we could go the other way, and actually set this value more generally,
and then pull it out of `buck2_data::RemoteCommand` (and move it up into
the `CommandExecutionDetails`). That would arguably be nicer, but it requires a
migration in Ingress.

If one day we want to log this, we can do so then.

Reviewed By: lmvasquezg

Differential Revision: D47754236

fbshipit-source-id: 2184faf5ef91bcaaab53a6c32ce3db1f59060928