buck2: capture more RE details in CommandExecutionKind
Summary:
We would like to capture more info when executing via Tpx, so we need to
include it here.

It feels a bit bad and overhead-ish to clone those Strings even though in the
common case we don't use them, but in the grand scheme of things, cloning a few
Strings probably doesn't really matter.

In the interest of actually bringing data to this conversation I did run a build of Adfinder 5 times with and without this.
The results appear to confirm it doesn't really seem to change much of anything:

 {F1057499394}

Note: I considered the possibility of Arc-ing those... but then you start synchronizing threads
instead of a temporary thread-local allocation, so that seemed not necessarily obviously better.

Reviewed By: themarwhal

Differential Revision: D47754237

fbshipit-source-id: 96958a58c82fe713c67968f0fec51e0db33829ff