buck2: nuke hang detector
Summary:
TL;DR: it's been a while now, without much use. Making it rely on spans is a bad idea. Removing it to not pay maintenance burden

Wrapping up here the conversations I've had with people on the team and my own thoughts about this. This is done mostly to organize my own thoughts, but maybe people find it interesting too

## 1. Context

[Hang detector](https://fb.workplace.com/groups/buck2dev/permalink/3419443435010371/) was intended as a quick experiment to detect deadlocks and gather debugging information. The idea is simple enough: every time the console shows 'nothing' for a long-ish period of time (1 min), trigger something that will gather necessary debug information to be able to tell what happened.

This approach didn't work for some reasons, I will dig deeper into them in a moment:
- Too many false positives
- Reportedly it would miss some deadlocks
- We needed more debugging information, specifically a threads dump.

**Too many false positives**

Turns out, we haven't been the most rigorous when adding spans to our commands. This is reasonable, there's hardly good incentives, or systems in place that would help the team have a correct and complete approach to the spans. Fixing this would be great, and some serious (and boring, and maybe not well rewarded) undertaking

This false positives would make the hang detector get triggered when there where things going on, leading to a high amount of false positives, resulting in ~30k unused triggers a day. And as [rajyengi warned us](https://fb.workplace.com/groups/buck2dev/posts/3419443435010371/?comment_id=3420697544884960&reply_comment_id=3420709014883813), this makes this useless

**It would miss some deadlocks**

Well deadlocks can happen while there's a span going on. That means, the console can be showing output, and still the program be completely blocked. This is completely missed by the hang detector.

**We needed more debugging information**

The debugging information required to effectively debug this information is costly to obtain. Costly in the sense that will heavily use system resources. This makes false positives even more concerning due to their detrimental impact on the user experience.

## 2. Could we just make it better?

Maybe we can come up with a better heuristic that will be more reliable. But I don't think that's possible. In the end, we will never avoid the fact that this is relying just on what the user can see (the spans). This is very limited, specially when it can't be relied upon. I think is better to rethink the approach, some ideas are already being thrown out in T158957272

## 3. Why remove it

We tried this "quick" experiment and it didn't work. Ergo is time to clean it up. There's little that can be of use elsewhere, and the extra code is a maintenance burden. The rare occasions when this has been useful, don't outweigh the extra added complexity this system brings.

## 4. What could have been done better

This project came up to my hands when I was arriving to the team, with little experience in the area. Looking back, there are a couple of things I would have done differently. First, set up expectations well. Make it clear from the get go this is an 'experiment' as this was, in the end, a relatively inexpensive bet on an idea. Secondly, set a time limit for this experiment to conclude. A month or 2 would probably have been enough.

**Wrapping up**

Correct me if I'm wrong on anything. I think I answered all of the team possible doubts. It will be a relief to delete this I think, it's something I've been wanting to do for a while now.

Reviewed By: krallin

Differential Revision: D47642265

fbshipit-source-id: c625e9c858ef9cb2c4c6c6dcb131787cee950c3f