Use weight = 5 rather than concurrency limit for JavaCD and KotlinCD worker actions
Summary:
This is a pretty imprecise science, so happy to discuss the best approach here.

Right, our JavaCD and KotlinCD workers have a concurrency limit, which is set to 16. We set the same limit regardless of the number of cores, so on M1s with 10 cores, the concurrency limit takes no effect, whereas on an 80 core devbig, we can have a bunch of actions waiting in the worker queue.

Using a `weight` instead feels like the right thing to do:
1) The number of worker actions running is now proportional to the number of cores.  These are pretty heavy actions, so for machines with fewer cores (i.e. laptops) it means that we won't saturate the machine with worker actions.
2) Less complexity - we just fit into the existing mechanism for deciding how many actions should run locally at any one time.

Needless to say, there might be tweaks that can be made to improve things, but until we have evidence that another approach is better, I think just using a `weight` is the best thing to do.

Reviewed By: cjhopman

Differential Revision: D47753011

fbshipit-source-id: 41805ea5b557673613c859febb23c397987c2dc5