Return ExitResult from a daemon connection error
Summary:
As the title says. This seems like a nicer implementation than D47343603.

Because of how this streaming function is set up, I can't reuse `gen_error_exit_code()` easily since the invocation recorder was already consumed during the failed buckd client instantiation. I could return the `EventsCtx` and use that to get the error cause, but that seems more complex than necessary. This is still adding a new variant to FailureExitCode, but at least we aren't treating all `Elapsed` errors the same.

Reviewed By: bobyangyf, krallin

Differential Revision: D47806133

fbshipit-source-id: 1b1b8b44a64450bc8d17b9e4a0b5282bddb28de7