Have Deferred::execute() take in a DiceComputations
Summary: The way I have it set up in D47729778, the async_scoped closure requires dice_ctx to be passed in, but we are also passing in deferred_ctx into the closure, so I can't borrow mutably twice with that setup. See D47795355 for passing in the dice_ctx via Deferred::execute() instead so the reference is not tied to the `deferred_ctx`.

Reviewed By: bobyangyf

Differential Revision: D47795355

fbshipit-source-id: 34b1bd427cc40d7b80b3e2df7ec8c83dff2c868f