buck2: add `read_parent_package_value` in `PACKAGE` contexts
Summary:
Like it says in the title, this allows you to call `read_parent_package_value`
to access the parent package. We don't call this `read_package_value` because
it'd be a bit confusing whether you are reading the parent's value or the
one you've just overridden (if you did).

That said, for convenience we suggest `read_parent_package_value` if the user
calls `read_package_value` from a Package context.

Reviewed By: stepancheg

Differential Revision: D47826462

fbshipit-source-id: 9713edcfc4151c829740226dc25b4dfb9ece74a6