buck2: buckconfig: parse config flags a bit earlier
Summary:
Like it says in the title. We don't need a cell resolver to be able to split
`cell//foo.bar=baz` into its components (`cell`, `foo`, `bar`, baz`), so let's
do that earlier.

My end goal is to catch folks setting `client.id` and to just not allow that.

Reviewed By: bobyangyf

Differential Revision: D47832943

fbshipit-source-id: 59d8c95225a0a5658447a292f1743702ad88fef7