chore(nix): use mold for linking inside 'nix develop' shell (#343)
Summary:
when using Linux, and developing with `nix develop` or `direnv`, go ahead and enable mold as the default linker.

We modify `RUSTFLAGS` in the shell environment to do this so that we don't have to fiddle with the internal `.cargo/config.toml` file; and because it's only for development anyway, this is strictly used only while developing, not while running `nix build`.

On my Ryzen 5600X, versus the default linker settings, this takes the link time for the (debug) buck2 executable from 32s to 3s.

Pull Request resolved: https://github.com/facebook/buck2/pull/343

Reviewed By: ndmitchell

Differential Revision: D47869265

Pulled By: themarwhal

fbshipit-source-id: 37f68fe1f4febcdca4f5b9c95e788b9c1f73e9e0