Be able to get a ProvidersLabel from a string
Summary:
Currently, it's possible to create a `ProvidersLabel` from a string, but it's quite clunky. You would have to call `unconfigured_targets()` on the string to get a `target_node`, then get the label from that, then pass it into `sub_target()` to get the unconfigured providers label.

To make life easier, let's add a function to convert a string to a providers label directly. This is especially useful for passing into/dealing with `exec_deps` and `toolchains` for configuring execution platform resolution, which both return a dict keyed by `ProvidersLabel`s.

Example usage:

```
def _impl(ctx):
    exec_dep = ctx.unconfigured_sub_targets("//foo:bar")
    bxl_actions = ctx.bxl_actions(exec_deps = [exec_dep], target_platform = "//foo:platform")
    dependency = bxl_actions.exec_deps[exec_dep]
    # run some actions using the exec dep
```

Reviewed By: bobyangyf

Differential Revision: D47403882

fbshipit-source-id: 642056937f72ac09ae80e036a990ef9fcfba7cad