buck2: move `allow_vpnless` parsing to the client
Summary:
Loading the var already happens here. Also have it parse it. This restricts
what we can do a bit (e.g. using `RolloutPercentage`), but:

- It allows the client to read it and that's important because the client *is*
  reading it.
- We don't use that extra flexibility currently.

Note that this doesn't change the value we end up using because the client
is already the one reading it (it just wasn't interpreting what it read).

Reviewed By: johansglock

Differential Revision: D47868908

fbshipit-source-id: 69b5930cb53a51ae71555a18eff795e6d82761fb