buck2: re log upload: extract some code
Summary:
This is going to make little sense as-is (I split it after the fact), but I
want to reuse this in the event log, so I'm extracting this.

Note how I spawn the reading, this is to ensure that if we spawn something
and use it (much) later, we still drain stderr. This matters a lot more in
the next diff.

Reviewed By: iguridi

Differential Revision: D47876119

fbshipit-source-id: 7e6c8fce550666abf2a5aa37f06c4afa59ae00cf