buck2: add a `allow_vpnless_for_logging` config
Summary:
I want to move all our corp logging to use vpnless to make it more reliable
(and namely: available when off-VPN).

I'm doing this because looking at our data quality alerts from corp, I
sometimes see us missing 100% of logs for 1 person for a little while, and that
feels like a off-VPN problem and not a transient issue.

This will allow that. I'm not just defaulting it to true because I want to
roll this out gradually, but I plan to take it to 100%.

Reviewed By: stepancheg

Differential Revision: D47913738

fbshipit-source-id: 5b127c803dd90c55c16cfa99a4ced542cf1fc3cb