Delay report of activation until we are sure we have fully executed the key
Summary:
We sometimes report activation twice, despite the logs showing the key was only ran once (and sometimes never ran due to it being a dependency check).

This is probably because we report activation twice despite the key only finishing once.

Specifically, modern dice has more exit points for cancellation. A key can be aborted / cancelled until it's state transitions to `cached`, in which case we guarantee that the task is considered finished and won't be ran again.
Until then, we can still cancel and re-check.

Reporting activation before we hit that point will result in multiple activations without actually having multiple activations. So, delay that reporting until we report `cached`

I think we can consider having the DiceTask State manage the activations itself down the line, but I'm going to be moving some things around with the DiceTaskHandle being updated by the state, so maybe do that after and see what it looks like.

Reviewed By: krallin

Differential Revision: D47939550

fbshipit-source-id: 2b6cd12e121bf25d3850ff37a0c390a4fda8d89e