buck2: event log: don't error out when you can't write the log
Summary:
That's really not a nice thing to do. If we can't write the logs, let's just
live with it and warn.

Note that we have alerts on data completeness to track cases where
we fail to upload logs. Those are more appropriate to track logging
failures than randomly failing user requests.

Reviewed By: lmvasquezg

Differential Revision: D47990736

fbshipit-source-id: 55f6fc1fcb30910e9f812834ee0b5e3ec6b40351