buck2: expose the action's scratch path via a dedicated variable
Summary:
Like it says in the title.

The intention here is to allow actions that are aware of Buck to opt into using
Buck's scratch path instead of TMPDIR.

Unlike TMPDIR, the scratch path makes stronger guarantees:

- It'll always be relative to your cwd (reproducibly)
- It'll always be exclusive to one action
  - Right now this is true in local execution
  - It's also true on our Linux RE
  - But it's not true on other RE platforms...

On the flip side, it's not created for you automatically, which is probably
fine, because if you're aware of this existing, then you can create the path.

This matters to help fix actions that are e.g. creating a temp file in TMP then
compiling it. Since the path to the file ends up in the object file (if only as
STT_FILE even in the absence of other debug info), you need the file to
have a deterministic name, but if TMPDIR may be shared with other actions, you
can't.

Having this relative scratch path makes it much easier to get this consistently
right.

I'm a bit on the fence regarding whether this and the TMPDIR should be the same
path. It feels a bit wrong to have two different paths. On the flip side, since
we use the scratch path as TMPDIR for local actions, this means the scratch
path gets created when running locally, but not when running remotely, which
feels like an awkward inconsistency. So, I went with putting the path we expose
as `BUCK_SCRATCH_PATH` *inside* the tmpdir, thus ensuring it's consistently not
there.

Reviewed By: wendy728

Differential Revision: D47991039

fbshipit-source-id: 2b4dfedbf4abc432528adb51f765fb900e4d83e7